Refactor: Extract workspace openssh-sftp-client-lowlevel
