+++ openssh-sftp-client-lowlevel/LICENSE
@@ -1,0 +1,1 @@
+../LICENSE
