+++ openssh-sftp-client-lowlevel/run_tests.sh
@@ -1,0 +1,1 @@
+../run_tests.sh
