+++ run_tests.sh
@@ -9,9 +9,13 @@ git submodule update --init --depth 1 --recursive
 export RUNTIME_DIR=${XDG_RUNTIME_DIR:-/tmp}
 
 # Remove all files in /tmp in the container
-mkdir -p $RUNTIME_DIR/openssh_sftp_client/
-rm -rf $RUNTIME_DIR/openssh_sftp_client/*
+mkdir -p "$RUNTIME_DIR"/openssh_sftp_client/
+rm -rf "$RUNTIME_DIR"/openssh_sftp_client/*
 
-cargo check "$@"
+for workspace in openssh-sftp-error openssh-sftp-client-lowlevel; do
+    cd "$workspace"
+    cargo test "$@" --all-features
+    cd ..
+done
 
 exec cargo test "$@" --all-features
