+++ src/lib.rs
@@ -1,5 +1,18 @@
 //! openssh sftp client, implements [sftp v3] accodring to
-//! [`openssh-portable/sftp-client.c`].
+//! [`openssh-portable/sftp-client.c`] and provides
+//! an easy-to-use highlevel API.
+//!
+//! All `async` functions in this module are cancel safe.
+//!
+//! Internally, this is archived by first writing requests into a write buffer
+//! containing [`bytes::Bytes`] and then flush all buffers at once periodically
+//! to archive cancel safety and improve efficiencies.
+//!
+//! However, cancelling the future does not actually has any effect,
+//! since the requests are sent regardless of the cancellation.
+//!
+//! Thus, if you cancel a future that changes the remote filesystem in any way,
+//! then the change would still happen regardless.
 //!
 //! ## Usage
 //!
@@ -34,8 +47,47 @@
 /// Changelog for this crate.
 pub mod changelog;
 
+use openssh_sftp_client_lowlevel as lowlevel;
 pub use openssh_sftp_error::Error;
-use openssh_sftp_error::{awaitable, openssh_sftp_protocol};
 
-pub mod highlevel;
-pub mod lowlevel;
+pub use lowlevel::{UnixTimeStamp, UnixTimeStampError};
+
+use bytes::BytesMut;
+
+mod sftp;
+pub use sftp::CancellationToken;
+pub use sftp::Sftp;
+
+mod cancel_utility;
+use cancel_utility::BoxedWaitForCancellationFuture;
+
+mod options;
+pub use options::SftpOptions;
+
+mod tasks;
+
+mod auxiliary;
+use auxiliary::Auxiliary;
+
+mod cache;
+use cache::WriteEndWithCachedId;
+
+mod handle;
+use handle::OwnedHandle;
+
+/// Module contains types for manipulating files.
+pub mod file;
+
+/// Module contains types for manipulating directories.
+pub mod fs;
+
+/// Module contains types for manipulating metadata of files or directories.
+pub mod metadata;
+
+type Buffer = BytesMut;
+
+type WriteEnd<W> = lowlevel::WriteEnd<W, Buffer, Auxiliary>;
+type ReadEnd<R, W> = lowlevel::ReadEnd<R, W, Buffer, Auxiliary>;
+type SharedData<W> = lowlevel::SharedData<W, Buffer, Auxiliary>;
+type Id = lowlevel::Id<Buffer>;
+type Data = lowlevel::Data<Buffer>;
